Made the l5 link show up in gazebo
According to
http://answers.gazebosim.org/question/20068/collada-model-is-not-visible-in-gazebo-89-if-transparent-a_one/

A certain tag with a certain attribute <transparent opaque="A_ONE">in a .DAE file does not show up in Gazebo 9 (but it does
in earlier gazebo versions).  The fix is to add
<transparency><float>1</float></transparency> to the DAE file after each
transparent "A_ONE" tag.

The other DAE files I checked in sawyer also uses this approach.

After adding these tags, the missing link displays properly in gazebo.